Refactor v2 trace collection helpers
